iip/requirements.txt
flask
waitress
werkzeug
pdf2image
pillow
pytesseract
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
filelock==3.12.0
fonttools==4.39.4
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
sympy==1.12
thop==0.1.1.post2209072238
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
ultralytics==8.0.115
urllib3==2.0.3
zipp==3.15.0